update assembled_sequence
This rule is now a checkpoint, because an unpredictable number of files was generated by the previous rule (getorganelle)
It is also an aggregate rule: with a list of inputs (expand(), no wildcards) and only one directory as output (no wildcard possible because not in input:). The code is quite similar that the "shell:" logic and the old rename_assembled logic, the different is that the script combines both logics

Note: I used a "script:" directive to show an alternative to "shell:" running only one script line.