Update snakefile
The get_assembled_samples method, is returning all FASTA created by the checkpoint, so {sample} is "reevaluted" and becomes a new list of sample IDs after the checkpoint is executed.

I created a new aggregate rule (expand(), no wildcard) to be able to used a function that call the checkpoint). The reason is because the checkpoint rule is also an aggregate rule so I cannot have wildcard in "checkpoints.X.get(**wildcards).Y", otherwise Snakemake will complain that the wildcard is not found in the "checkpoint wildcard". I don't know how to make it work without this last aggregate rule, maybe this is a better solution. Hopefully you will use a rule like this to generate a report

This pattern used the second example (second large code block) described here https://snakemake.readthedocs.io/en/stable/snakefiles/rules.html#data-dependent-conditional-execution